feat: adds notebooks-api feature flag
+ refactors API calls into their own file
